// packages/omo-central-webnative/src/types.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type AddCirclesTokenTransferInput = {
  createdAt: Scalars['String'];
  createdInBlockNo: Scalars['Int'];
  createdInBlockHash: Scalars['String'];
  subjectAddress: Scalars['String'];
  predicate: CirclesTokenTransferPredicate;
  objectAddress: Scalars['String'];
  value: Scalars['String'];
};

export enum CirclesTokenTransferPredicate {
  GivingTo = 'GIVING_TO',
  ReceivingFrom = 'RECEIVING_FROM'
}
// ...
export type MutationAddCirclesTokenTransferArgs = {
  data: AddCirclesTokenTransferInput;
};